Remove two capability types
- tosca.capabilities.nfv.Forwarder
- tosca.capabilities.nfv.VirtualLinkable

They are not present in csd04.
The changelog at the end of csd04 states:
"Deleted NSD related content, v1.0 will only cover VNFD model."
The second capability was a part of the NSD model.
The first capability was a part of the VNFFGD model, which probably
also got deleted, since it is not the VNFD model.